Changed Haddon Utility to fit the recent changes in timer system in 1.3.1, and added weak adaptation to the lack of timer accumulator. Therefore, the Utility Implementation is extended to a ModLoader version (HaddonUtilityModLoader) instanciated in HaddonBridgeModLoader

+++ mc-src/eu/ha3/mc/haddon/Utility.java
@@ -112,7 +112,7 @@ public interface Utility
 	
 	public void closeCurrentScreen();
 	
-	public int getClientTick();
+	public long getClientTick();
 	
 	public void printChat(Object... args);
